perf: vectorize loops, eliminate float64 waste, reduce allocations
- Vectorize generate_inverse_distortion_map: replace O(H*W) Python loop
  with np.mgrid + fancy indexing scatter-update (4 iterations)
- Vectorize create_piecewise_affine_maps: replace O(H*W*N) nested loops
  with np.einsum IDW interpolation over full grid
- Vectorize perspective_bboxes HBB min/max: np.column_stack + axis min/max
  replaces per-box list comprehension
- Vectorize solarize LUT: np.where replaces list comprehension
- Vectorize ordered_dither_uint8 LUT: np.minimum + fancy indexing
- Vectorize GridDistortion RNG: single np.random.Generator.uniform call
- Fuse sharpen_gaussian: add_weighted replaces img + alpha*(img - blurred)
- Optimize to_gray_desaturation uint8: uint16 max+min with bit-shift
- Optimize slic: np.where, meshgrid centers, precomputed inv_grid_step_sq
- Preallocate mask in distort_image outside loop, reset with [:] = 0
- Preallocate poly_mask in add_shadow outside loop
- In-place np.clip with out= in dithering, bbox clipping, hsv adjustment
- Fix float64 waste: np.float32 dtype on np.arange, np.linspace, np.zeros
  in meshgrid, control points, gaussian kernel, contrast LUT
- Use snow_texture[:,:,np.newaxis] instead of np.dstack([arr]*3)
- Use first_result[np.newaxis] instead of np.array([first_result])
- Guard np.ascontiguousarray with C_CONTIGUOUS flag check
- Replace cv2.LUT with sz_lut in dithering_functional
- Remove dead code: ensure_contiguous_output, find_keypoint
- Remove duplicate num_boxes assignment in mask_to_bboxes
- Simplify compute_affine_warp_output_shape type handling
- Simplify _equalize_pil histogram filter
- Update coding_guidelines.md with performance optimization checklist
- Update CLAUDE.md with expanded anti-patterns list
- Update benchmark skill with JSON output/comparison workflow

Made-with: Cursor

+++ albumentations/augmentations/blur/functional.py
@@ -390,7 +390,7 @@ def create_gaussian_kernel_1d(sigma: float, ksize: int = 0) -> np.ndarray:
     # Guard against sigma=0 (would cause division by zero)
     if sigma == 0:
         # Return identity (delta) kernel when sigma is zero: no blur is applied
-        kernel_1d = np.zeros(size, dtype=np.float64)
+        kernel_1d = np.zeros(size, dtype=np.float32)
         kernel_1d[size // 2] = 1.0
         return kernel_1d
